refactor(agents): build agentbox opencode from the command factory
`commands/opencode.ts` (1,327 lines) becomes `agents/opencode/runtime.ts` plus a
descriptor. Its three real differences from codex are now declarations rather
than a forked file: no skip-permissions flag (`skipPermissions: null`), no seed
prompt (`acceptsSeedPrompt: false`, which is what routes a resync warning to
stderr), and a login that always needs a TTY because its provider picker
prompts.

It needs no teleport hook at all: `caps.teleport: 'stub'` already makes
`prepareTeleport` refuse with the reason the registry row carries, so `-c` and
`--resume` behave exactly as before with no per-agent code.

Golden CLI-surface fixture unchanged.
